chore(tracing-otel): inline HeadersTextMap{Getter,Setter} into OtelTracer
Both adapters were package-private SPI helpers with exactly one production
callsite (OtelTracer itself), zero tests, zero examples. Their Javadoc already
declared them as "Used by OtelTracer#..." — conceptually private. Pulling them
into OtelTracer as private static nested final classes (preserving the
stateless-singleton INSTANCE shape) drops two top-level files, keeps the
TextMapGetter/Setter SPI shape intact, and tightens encapsulation: the
adapters now cannot be referenced from anywhere else in the module.

Net LOC: -61 / +53 = -8 lines, one less compilation unit.

JPMS unchanged: package-info still exports the same package; the deleted
classes were never independently exported, so no module-info edit needed.
